requirements.txt
argon2-cffi==20.1.0
cffi==1.14.5
click==7.1.2
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pycparser==2.20
six==1.16.0
SQLAlchemy==1.4.15
Werkzeug==1.0.1
